#9 Refactor screen and fix typo
