+++ lib/presentation/groups/group_item/group_item_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:split_the_bill/domain/group/states/group_state.dart';
-import 'package:split_the_bill/presentation/groups/group_item/group_item_silver_list.dart';
+import 'package:split_the_bill/presentation/groups/group_item/group_item_sliver_list.dart';
 
 class GroupItemScreen extends ConsumerWidget {
   const GroupItemScreen({super.key, required this.groupId});
@@ -18,7 +18,7 @@ class GroupItemScreen extends ConsumerWidget {
       ),
       body: CustomScrollView(
         slivers: [
-          GroupItemSilverList(
+          GroupItemSliverList(
               scrollController: scrollController, groupId: groupId),
         ],
       ),
