+++ lib/presentation/groups/group_item/group_item_sliver_list.dart
@@ -6,8 +6,8 @@ import 'package:split_the_bill/presentation/groups/group_item/controllers.dart';
 import 'package:split_the_bill/presentation/shared/async_value_widget.dart';
 import 'package:split_the_bill/presentation/shared/bill/bill_list_tile.dart';
 
-class GroupItemSilverList extends ConsumerWidget {
-  const GroupItemSilverList(
+class GroupItemSliverList extends ConsumerWidget {
+  const GroupItemSliverList(
       {super.key, required this.scrollController, required this.groupId});
   final ScrollController scrollController;
   final String groupId;
