+++ lib/presentation/groups/groups/groups_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:split_the_bill/presentation/groups/groups/add_new_group_button.dart';
+import 'package:split_the_bill/presentation/groups/groups/groups_sliver_list.dart';
 
-import 'groups_silver_app_bar.dart';
-import 'groups_silver_list.dart';
+import 'groups_sliver_app_bar.dart';
 
 class GroupsScreen extends StatelessWidget {
   const GroupsScreen({super.key});
@@ -14,9 +14,9 @@ class GroupsScreen extends StatelessWidget {
     return Scaffold(
       body: CustomScrollView(
         slivers: [
-          const GroupSilverAppBar(),
+          const GroupSliverAppBar(),
           const AddNewGroupButton(),
-          GroupsSilverList(scrollController: scrollController),
+          GroupsSliverList(scrollController: scrollController),
         ],
       ),
     );
